add part4 exercise 4.21*

+++ part4/Bloglist/controller/blogs.js
@@ -48,9 +48,26 @@ blogRouter.post('/', async (request, response) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
+  if (!request.token) {
+    return response.status(401).json({ error: 'token missing' })
+  }
 
-  response.status(204).end()
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const user = await User.findById(decodedToken.id)
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() === user.id) {
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  } else {
+    response
+      .status(403)
+      .json({ error: 'User is not authorized to delete this blog' })
+  }
 })
 
 blogRouter.put('/:id', async (request, response) => {
